chore(code-garden): split BookmarksTab.jsx into focused sub-components [2026-W30]
Non-functional change from repo audit 2026-W30.
Finding: BookmarksTab.jsx monolith (811 lines) — extract Sankey + time-series
chart + funnel/topics + pending approvals into dedicated sub-components so the
container stays under 200 lines.

Container now owns state, data loading, and memoized derivations only; each
extracted component owns its own render + local state. All data-testid
selectors preserved, so BookmarksTab.test.jsx and the wider mission-control
test suite continue to pass (175/175).

Files: 9 tabs/* files (1 refactor + 8 new) + 1 audit marker.
