+++ docs/repo-audits/2026-W30.md
@@ -347,7 +347,7 @@ No new performance concerns this week. The Content Scheduler auto-post correctly
 - *Effort:* M
 - *Follow-up classification:* `garden-safe` (behaviour-preserving refactor).
 
-**2-B. Extract `BookmarksTab.jsx` Sankey + time-series into sub-components**
+**2-B. Extract `BookmarksTab.jsx` Sankey + time-series into sub-components** ✅ [PR #363](https://github.com/Stoffer-Industries/sindustries/pull/363)
 - *Description:* Split `BookmarksTab.jsx` (811 lines) into `BookmarksTab.jsx` (container), `BookmarksSankeyChart.jsx`, `BookmarksStatesOverTimeChart.jsx`, `BookmarksFunnelList.jsx`, `BookmarksPendingApprovals.jsx`.
 - *AC:* Container ≤ 200 lines; each chart component owns its own render + local state.
 - *Effort:* M
